Move UnknownSiteException to astropy.coordinates.errors
