+++ astropy/coordinates/earth.py
@@ -9,6 +9,7 @@ from ..extern import six
 from ..utils.exceptions import AstropyUserWarning
 from . import Longitude, Latitude
 from .builtin_frames import ITRS
+from .errors import UnknownSiteException
 
 try:
     # Not guaranteed available at setup time.
@@ -208,7 +209,6 @@ class EarthLocation(u.Quantity):
         --------
         get_site_names : the list of sites that this function can access
         """
-        from .sites import UnknownSiteException
         registry = cls._get_site_registry()
         try:
             el = registry[site_name]
