+++ astropy/coordinates/errors.py
@@ -163,3 +163,14 @@ class ConvertError(Exception):
     """
     Raised if a coordinate system cannot be converted to another
     """
+
+
+class UnknownSiteException(KeyError):
+    def __init__(self, site, attribute, close_names=None):
+        message = "Site '{0}' not in database. Use {1} to see available sites.".format(site, attribute)
+        if close_names:
+            message += " Did you mean one of: '{0}'?'".format("', '".join(close_names))
+        self.site = site
+        self.attribute = attribute
+        self.close_names = close_names
+        return super(UnknownSiteException, self).__init__(message)
